Extract the scalar differential operators to their own file
Pdop and Sdop know almost nothing about geometric algebra, so there is no need for them to be in the `mv.py` file.

It's likely in a future commit that `Dop` can be made a subclass of `Sdop`, with thin layers of `Mv` wrapping.
