+++ galgebra/__init__.py
@@ -16,6 +16,7 @@ Submodules
     ga
     mv
     lt
+    dop
     printer
     utils
 """
